requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828585465/work
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
audioread @ file:///D:/bld/audioread_1607457690506/work
backcall==0.2.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
cachetools==4.2.1
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600699246375/work
chardet==3.0.4
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
configparser==5.0.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
entrypoints==0.3
ffmpeg-python==0.2.0
fsspec==0.8.5
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598865816163/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
jedi @ file:///C:/ci/jedi_1611315808330/work
Jinja2==2.11.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1607956439537/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kiwisolver @ file:///D:/bld/kiwisolver_1610099948875/work
librosa==0.6.3
llvmlite==0.35.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib @ file:///D:/bld/matplotlib-suite_1605180506412/work
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
musdb==0.3.1
museval==0.3.0
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914925608/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
notebook @ file:///C:/ci/notebook_1601501643625/work
numba @ file:///D:/bld/numba_1607010604208/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging==20.4
pandas==1.2.1
pandocfilters==1.4.2
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pathtools==0.1.2
pickleshare==0.7.5
Pillow @ file:///D:/bld/pillow_1610407616132/work
prometheus-client==0.8.0
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.14.0
psutil==5.8.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
python-dateutil==2.8.1
pytorch-lightning==1.1.6
pytz==2020.5
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7
scikit-learn @ file:///D:/bld/scikit-learn_1611079981849/work
scipy @ file:///C:/bld/scipy_1609458016927/work
Send2Trash==1.5.0
sentry-sdk==0.19.5
shortuuid==1.0.1
simplejson==3.17.2
six @ file:///C:/ci/six_1605187374963/work
smmap==3.0.5
SoundFile==0.10.3.post1
stempeg==0.2.2
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
torch==1.7.1
tornado @ file:///D:/bld/tornado_1610094858390/work
tqdm==4.56.0
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
urllib3==1.26.3
wandb==0.10.15
watchdog==0.10.4
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
yarl==1.6.3
zipp @ file:///tmp/build/80754af9/zipp_1603306325174/work

numpy~=1.19.2
wget~=3.2
hydra-core==1.1.0